Bug Fix 1262
Previously printBody() in page_poll.inc returned false if the question is not selected or the selected question is not live, which
prevented from previewing the site.

Now it does not check the status of the question asset to show on the frontend. To hide the question, clear the current question option
in the details screen.

In addition, the poll question is linked under poll page dependently, i.e. it's meaningless to check the status of the poll question to
display on the frontend since it changes with its parent asset(page_poll).