Mysterious PHP Notice error fixed.
When the user edit multiple plots in the graph, the index of the current editing plot is passed through $_POST var. Then, each display
function looks up the variable to check whether the plot is deleted or not. However, when the page is loaded at the first time, the var
was empty so that it generated an undefined index notice. Then... if the user moves to a different plot, it dissapears.

Now, it's fixed.