Bug fix for datetime field entries in submission logs and on receipt emails.
All the entered date becomes '01-01-1970'.

The problem was caused by mktime() function which can only generate UNIX timestamp between the Unix Epoch and the time specified. The year can be
handled differently or incorrectly depending on the operating system or PHP version.

Now, when we convert datetime format, we don't pass year to mktime() funciton.