The beforeAssetDeleted trigger event is now fired by the asset manager any time it is about to bring an asset into the state where it is linked only under the trash.  This is done in the deleteAssetLink function.
An extra transaction has been added to the trashAsset function to make sure the whole operation is rolled back if the trigger (which
is fired when the final non-trash link is about to be deleted) fails.

The 'moving' parameter of deleteAssetLink has been removed as it was no longer necessary and caused confusion.  It was added in the
brief period when the old link was deleted before the new one was created, but this is no longer the case.