Added a new configuration screen - 'Apache Configuration'. This screen allows the configuration of Apache related settings; currently the user can configure the names of certain request notes that are set by MySource Matrix.
A request note is similar to an environment variable - by using the apache_note() function we can make values accessible to Apache and/or its Modules. With this change, the user can define the name of the Apache request note for both user id and user name which can then be used for things such as logging.

NOTE: If you don't run step03.php, the configuration file for this new screen won't be created in data/private/conf/, the necessary constants won't get defined and you'll get nice little error messages on every page/screen :)