Bug Fix #1307 Import Tool HIPO dies when HTML tidy option is chosen
Now, a temporary HTML file is created to run tidy on it, instead piping the HTML content to tidy.

Also, some Microsoft Word tags have been added to clean up.

Note: I just found out that the result of doc2html.exe is dependant on the version of Microsoft Office installation. This could be
potential annoyance for developers later.