Introduction of the Mass-Clone tool.  This tool allows many clones of an asset to be made in one go.
The tools interface (and the existing tools) have been modified so that the commit button of the
interface is painted centrally rather than by each tool, and the processTool function returns
a boolean indicating success or failure (failure triggers the re-painting of the interface).

The clone-assets HIPO has been altered slightly so as to be able to accept and act upon a
'number of clones' parameter for each of the clonee assets.  It can also be given a 'finalise_function'
parameter which is a callback to a (static) function that will be called after each new clone
asset has been created.  This will allow subclasses of the mass-clone tool to individualise the
clone assets, for example by giving sequential values to their datetime attributes.

Step 3 and compile_locale required.