made some significant changes to the way hipo_job_edit_status works. previously this hipo could edit the status of assets in two ways:
1. single asset - changes status of asset in question and any dependants of that asset
2. multiple asset (cascade) - changes status of asset in question and any children (both dependant and independant) of that asset

the problem with the way this worked is that for any asset with dependants, _updated() would be called before the dependants had been processed. for example, a rebuild cache trigger fired during _updated() of an under construction during status change from to live would cache dependants that had not yet been processed (and therefore still under construction).

to correct this the hipo_job_edit_status will now only process independant assets. if during this processing it comes across an asset with dependants it will change it's status without calling _updated() and will proceed to spawn another hipo_job_edit_status to be run inline. this inline hipo job will process the dependants of the asset only. once this spawned hipo is complete, it will then execute _updated() on the parent of those dependants.