Bug Fix #1345 paint() function does not construct valid file path
When paint() function can not find the design file for the design customisation, it should look for the parent design or design customisation to locate the design file. Previously, it could not handle the customisation of the customisation. Now, it loops until it meets the design file, or not it reports an error.