Added a new system integrity script to report on database foreign key integrity. This script does not attempt to fix the integrity of the database, rather it reports on the number of invalid records and the values of the foreign keys. This information could then be used by the user to write an SQL query to remove the invalid records if desired.
It currently checks all of the sq_ast_* tables and adding new tables/foreign keys for checking is simple. If you come up with a better way of displaying this information via the command line please make the change (I've tried to keep it simple, but it could get confusing on systems with large asset counts).