Bug fix $1438 Having a form input name="submit" prevents "commiting" the page contents
Since Matrix commits the form in the backend or limbo using sumbit() function in javascript. If the content type is WYSYWYG, the input tag itself is printed for the preview mode. When the input tag's name attribute has the value of "submit", it overwrites the submit function in form object as both of them can be referred as "form.submit".

Previously, the backend or limbo interface freezes since the javascript disables all the buttons before it calls the submit function, which resulted in javascript error, so that the user can't take any further action except deleting the asset.

Now, if the input field contains the name, "submit", it will pop up error message and stop submitting the form.