New feature has been added to WYSIWYG
The user can choose the option, 'Disable Image Resizing' in the WYSIWYG section from  Global Preferences.

Once this option is set to yes, the users can not change the size of the image in the WYSIWYG as long as the image is from the Matrix's Image Asset. The users still can attempt to change the size by dragging the image's edges in the design mode or directly modifying the source code, however, Matrix Insert Image plugin will reset the original size back.