#1462 Cascading roles and permissions gets mixed up
'Cascade Permission' option got the list of permissions from the permissions view instead of permissions table only. The problem occurs when the parent asset is assigned to Role since it cascades the permission for the user who is assigned the Role as well.

Now it gets the list of permissions with the collapse roles option as TRUE, so that it does not cascade to the list of users who are assigned to the Role.