#1463  	Search Folder Bug
The asset map generated an error when the user tries to expand the children of the search result in the search folder.

The users still can access to the search result assets but they can not expand their children since there is no real link between the search folder and the search result.

getAssetMapLinks() now always return the number of kids as zero, so that the user can not have an expanding button on the asset map. It also returns the shadow assetid for the search result in order to intercept the refresh request at the search folder.

getAsset() will just return the real asset object, and getLinks() will always return 0 when the links from the shadow asset is requested. This iwll effectively block the user from expanding the asset map.