Re-structuring the data directories.
There have been a couple of issues in Matrix, which are related to the inode limit in the file system. Basically, one directory can have the maximum 32,000 files or sub-directories unless you change the setting during the OS installation. The number seems big enough to handle most of applications, but the problem occurs in Matrix as it organizes assets in the data directory according to the asset types. Since all the same type of assets are stored in one directory, for example, /data/private/assets/word_doc, some websites with more than 32,000 assets of the same type will run out of inodes.

In order to overcome this limitation, the way of storing assets in data directory has been changed. Instead of having assetid folders right under asset type folder, it will create one more level down with the hashed assetid.

1. SQ_CONF_NUM_DATA_DIR contant has been added to main.inc. The default is set to 20 so that each asset type folder can have maxinum 20 subfolder to organize the assetid folders.

2. asset_data_path_suffix() function in /core/general.inc has been updated.


The new script, upgrade_data_dirs.php will attempt to update all the directories and the lookup URLs. The following is the recommended steps to run the script.

1. Run cvs up.

2. Run the following command on the command line. It will show you the number of expected directories and database rows that  the script will try to update.
$ php ./scripts/upgrade_data_dirs.php -s . --report-only

3. Run the following command to start the script
$ php ./scripts/upgrade_data_dirs.php -s . --start-upgrade

You can optionally add --show-details to see the detailed progress as well. Once the script finishes, it will show the result table. If it says, "Congratulations", everything went good. If it tells any error messages, the expected actions couldn't be finished all. You can restart the script without problems to see which steps are failing.

NOTE : This script runs independently from Matrix core system. It means it manually updates the file system and database rows without using Matrix functions like HIPO. In other words, make sure that nobody is updating the system while you are running the script. This is mainly due to the performance reasons of the script.

After the script is finished, we recommend to clean up the existing cached pages if there are any since the webpaths to the public directory might have been changed.