Some changes in the way of sending the file to the frontend.
readfile() function in PHP occasionally fails to read the end of file for paticular types of pdf files.

Instead of using the wrapper function for all types of files, the child of file asset now can override the way it prints on the frontend in _printFronendFile().

Now, pdf_file asset sends the file by manually looping buffer by buffer.