Bug Fix #1482 Categroy Keyword on the Default Search Bodycopy
The 'category_chooser' dropdown keyword was not properly replaced in some of bodycopies under Search Forms. Now, the category chooser will appear for the search-in-search and cutomised form bodycopies.

By the way, the available keywords in the backend for the search page are not consistent with the current search package manual. The manual is being updated now, and further fixes will be made in search page.