File downloading problem in the frontend.
The previous commit was only applicable to pdf_file type of assets, which was failed to be downloaded the full size of itself. Now, the same problem comes from some video files as well and it strongly implies the problems with the wrapper funtion we are using, readfile().

Now, instead of calling readfile(), all the file will be served by manually looping with fread() function until it reaches to the end by default. If you find any issues with this implementation, let me know.

At the moment, it echoes 8kb per loop. We may need to find the optimal size for this in the consideration of performance when the file size is so big.