Bug Fix #1524.
getLineageFromURL() function did not check the asset status, esp Safe Edit, and returned the lineage information directly from database. It resulted in displaying the modified asset's short_name while it's in Safe Edit mode in the breadcrumb.

The fix has been made so that it loops through each assets in the lineage and if it is in Safe Edit mode, it gets name and short_name from the saved system version.