Now, the search index performs the duplicated check before it runs INSERT query.
Since the Search Manager extracts a raw number from text value, it can have duplicated values. For instance, if there are texts in the content like, '50000' and '50,000', then both of them will be a string, '50000' to be indexed resulting in duplicated key violation error in db.

The fixed code will check the hash of the key and if it finds any duplications, then it will skip.