Buf Fix#1541 Cannot display 'nested content' that comes from 'Type Format'->'Default Format' of Paint Layout.
Paintlayout bodycopy is not updated if the content type is updated but its attributes were not changed. For instance, the nested content type, it does not change any of its attributes although you change the asset to nest within and it changes the content of the page. Consequently, saveAttributes() function can not propagate to all the parents resulting in paint layout bodycopy unchanged. This problem can be applied to any paint layout bodycopy.