This is the first commit of the export_to_xml.php script.
This script takes a few arguments:

# php scripts/import/export_to_xml.php /path/to/matrix <start_asset_id> <destination_parent> <destination_link_type> > export.xml

The last two arguments may disappear later. :)

Essentially, the script does the following:

Starting at the <start_asset_id> it writes out the XML used by the import_from_xml.php script for each asset it finds. It then recurses through
dependant and non-dependant children, so it will export the entire hierarchy from the <start_asset_id> down.

It exports:
   - asset with all attributes
   - Bodycopy content (including remapping links)
   - Links (including linked assets)
   - Permissions

It has a bit (but not much) intelligence:

For each remap, it determines whether the source asset is being recreated during the import. If so, it sets the target asset as the remap, instead of
the source. For example, if I have a standard page with a nested bodycopy that's nesting another standard page, the exporter will determine whether or
not the nested standard page is being recreated. If it is, it'll set the nested bodycopy to point to the NEW asset ID. If its not, it'll point it to
the OLD asset ID. Therefore, if you want to export a site from one Matrix system to another, please ensure that you don't have links from the site
you're exporting to assets you're not exporting.

There is a high likelihood that the script will create invalid XML. Warranty void where permitted.