BUG FIX #1567: Despite all the work including prefixes and what-not, there's no way that more than one option list attribute could be printed per screen and still maintain 100% functionality.
The option list attribute would paint it's elements inside a div, and then print some inline javascript that would set the value of 'optionList' to the div that was just printed. Because of this, 'optionList' would always refer to the last option list attribute painted on the screen.

Javascript functions listMoveUp() and listMoveDown() would then use this global 'optionList' variable, meaning that only the last option list attribute printed could have it's elements moved around.