Fixing bug #1594 (disabling css file caching didn't work)
The CSS File (aka CSS design) asset now creates a file asset to act as its cache as soon
as it's created.  If the cache_css attr is set to true it updates the content of the
cache file asset every time the parse file is changed.  An upgrade function has been
added to make sure that all existing CSS File assets now have a cache file (even if it
is currently unused).

The code that the linked_css design area writes to the design file now checks if caching
is enabled on the selected css file asset.  (This is a slight performance cost but is
necessary for correctness).  Iff caching is enabled and the cache file asset has a URL,
that URL will be used.  Otherwise the URL of the actual css asset is used.

The interface of CSS File has been improved slightly to make it clearer how this all works,
and to remove the irrelevant "leave empty" option.

Step 3, compile_locale and reparse of any designs using the linked_css design area required.