i#1601 Nested Content recursion errors
When the asset listing is printed as a nested content in the paint layout, and %asset_contents% keyword is used to print the asset, even though the dynamic root node is used for the listing engine, the nested content type thinks the same asset is printed recursively and generate the warning message in the content file.

It was caused since only the assetid is being used for tagging for the indication that the asset is already printed.

Now, if the nested content asset has _getCacheKey() function, the cache key is used for tagging instead of assetid so that it can differentiate the same asset listing with the dynamic root node.