Bug Fix #1620 Custom Form causes spurious workflow errors
The current workflow manager does not handle the shadow assetids, and the assetid becomes the same as the bridge assetid when the SQL is constructed. Consequently, the asset status thinks the form question is under construction and the workflow is running at the same time, so it generates the error messages.

The solution to this problem is to quitely mute processStatusChange() function call for Form Questions since it gets its status from the bridge asset. This bug can potentially affect the asset types which implements the bridge with the workflow applied.