Bug #1659 Workflow doesn't cascade to dependent assets when Cascade is unticked
When the cascade option is unticked, the workflow is not applied to the dependent children. The workflow should be cascaded to the dependent children regardless of cascade option.

Also, freestyle() function in edit workflow schemas hipo job did not acquire locks properly resulting in failing to update workflow. Fixed now.