Bug #1686 The asset map fails to locate any url of the site under the folder during the initialisation
When the asset_map.inc sets the initial selection for the asset map, it calls getLineageFromURL() in asset manager, and it only returns the lineage up to the top site asset. If the site is nested under any other type of assets, the value of the parameter passed into the asset map does not include the lineage above the site which results in failure message.

Now, _getInitialLineageSelection() function checks whether the asset exists between the site and the root, then if so, it fixes the parameter string before it passes.