Bug Fix #1700
When multiple root nodes are chosen for search page, %asset_lineage%, and %asset_lineage_linked% keyword do not display the root node correctly for search results not belonging to the top root node. It was caused because it used to find the root node by calling getLink() instead of getLinks() considering multiple root nodes. It's been fixed.