Bug Fix #1708 img tag code generated from wysiwyg shows non-compliant border attribute
The current WYSIWYG insert image plugin adds an image tag with border attribute with is not compliant to HTML 4.01 Strict. This fix will change the tag into style attribute. For example, border="1" will be style="border: 1px solid rgb(0, 0, 0);"

The existing image tags will not be fixed automatically. Those images are recommended to be opened with the plugin again and committed so that the plugin can apply the changes.