New Trigger Actions have been added to IPB package
- Trigger_Action_Login_Ipb
- Trigger_Action_Logout_Ipb

They should be used in conjunction with Trigger_Event_User_Login and Trigger_Event_User_Logout events to get the username and password. If the same username/password exist in IPB, it automatically logs the user in or out from IPB. These actions require IPB SDK installation on the server side. It is recommended to choose the right version of IPB SDK which is compatible with the installed IPB. For more information, please visit http://ipbsdk.sourceforge.net

Note : This package has been tested with IPB 2.0.x and 2.1.x together with IPBSDK 1.5.0 and 1.6.x (beta test) versions respectively.