Bug Fix #1723 Insert Anchor Tag Plugin does not work without Visual Aid Plugin installed.
Matrix Insert Anchor and Choose Language plugins insert different HTML if the Visual Aid mode is on in WYSIWYG. If the Visual Aid mode is on, the global variable 'visual_aid_mode' is set to TRUE, and other plugins can refer it to change the inserted HTML.

However, if the visual aid is not installed, the variable type becomes 'undefined' (null) and it was not converted to boolean false correctly causing undefined variable javascript error. Consequently, the user could not insert anchor tag, or even close the pop-up plugin window.

It's been fixed now.