Bug #1734 - "WYSIWYG Editor does not work in IE"
WYSIWYG replace_entities() function has been updated.

Previously, this function includes the non-ASCII encoded characters in order to replace them with HTML entities. If the encoding type of the  wysiwyg.inc is not matched with the system default character set type, IE complains that invalid characters are found and javascript does not work at all, i.e. the user can not commit the page.

In order to fix this, all the non-ASCII characters are removed from wysiwyg.inc, and represented as a numeric character code.