Bulk File Import Tool.
This tool provides the backend interface of import_files.inc script.

The user should create a directory on the same server for file import and set the directory path in the Import Tool Manager. They can choose the link type and allow_unrestricted option during the bulk import.

Any directories under the import directory will be listed in the table. If there are any files in the import directory, they can be imported at once.

Note: It only handles folders one level down. Do not nest the directories.