BUGFIX #1864: User configuration of the Messaging Service using the MySource Matrix backend interface would cause message sending to break.
This was caused by the use of a textarea user input for management of the white/black lists for messages (exploding the value on '\n' when, depending on the backend encoding, the newlines would not always be '\n'). Fixed by changing the input method for the white/black lists to use the interface of the option list attribute.