Backend search: the jewel of the crown :-)
When search package is installed, the Quick Search tool allows you to enter keywords and return assets that match those keywords. Results are displayed 5 at a time by default (changeable via Search Manager preference) and gives you details of where it was found, as well as a tag line (ie. more info, Find Me, and details screen link). In Oracle, the advanced Oracle Text searching is also available.

If a match is found both on asset ID/URL and keywords, both will be displayed in the one window. If a match is found on asset ID/URL only, then behaviour will be as is currently - ie. a straight jump to that page.

Improvements to the header frame include waiting for all messages to pass through before updating it (otherwise message frames would be invalidated) - but forcing an update one minute before locks are lost. The messages now slide in quicker, making long messages like search results more bearable to wait for.

A full system upgrade (at least step_03/compile_locale, I think) are required before this will work completely, due to the need for Search Manager preferences and new translations.