Bug Fix #1875 Files not being deleted when new file is uploaded over it
The file type of asset fails to clear up the old checked-out file from the data directory if the user changes the name of file and uploads a new file at the same time. Consequently the old file still exists, for instance, in the public data directory, and remains accessible.

This problem is due to the fact that the name of the file is updated first during the process phase, then uploading file updates the file name  again. When the name attribute is about to be saved, the new file name in the file name field becomes the old file name and the uploaded file name becomes the new one. At the end, the original old file existing in the file system is not deleted.

The fix prevents the file name from being processed if the user is uploading the file at the same time. We might be able to consider some changes in the future so that when the user types the name and upload the file together, the newly created file will have the new name.