Bug Fix #1897 Malformed/invalid HTTP header response causes link report to break
Some malformed/invalid HTTP header can contain <script> tag, then when the link report page displays the report in the backend, the printed HTML includes open <script> tag without closing it. Internet Explorer treats the rest of HTML as script, breaking the backend. The solution to this is to encode html entities so that whatever response matrix got it from the URL, it displays the result properly without breaking HTML.