Bug #1900 Thesaurus XML import broken with Squiz Server
Process function of 'Erasing Thesaurus Contents' step in Import Thesaurus XML hipo job did not have form submission check before. This results in skipping the confirmation screen of the hipo job if the Squiz Server is on.

This bug is only discoverable if HIPO is using Squiz Server. Since the Hipo Job processes the current step first and tries to paint the web interface if there is any.  After Hipo Herder changes the running mode from 'server'  to 'web' for  'Erasing Thesaurus Contents' step, the process function is called again, then it erases the thesaurus terms and steps moves to the next skipping the confirmation phase.

By adding the form checking, hipo job stay in the same step until the confirmation screen is painted and the user clicks 'Next' button.