New features for asset builder and, by extension, account manager.
You can now select some metadata schemas that you want to apply to the created asset,
and allow the frontend user to supply values for their fields when creating the asset.

By default, the interface to supply values for the schemas' fields is printed at the
top of the normal create form for each asset type.

However, it is also now possible to customise the layout of the create form
shown on the frontend by asset builder (even if not using metadata).
On the details screen you can enable create form customisation for
one or more asset types, whereupon a layout asset for that asset type
is created under the asset builder.  You can include limbo keywords for
any section or field from the details or metadata screens in the layout,
thus configuring the display of the %create_form% keyword on the frontend.

Step 3 and compile_locale required.