Fulfilling  Feature #1744 by adding a new File Preference option.
Now the site administrator can restrict the maximum size of pre-uploaded file when file type of asset is created. vaildFile() function in file.inc will use SQ_PREUPLOADED_FILE_MAX_SIZE instead when the file is created from the pre-uploaded file in order to prevent PHP max file size blocks the action.

Note that it does not mean that you can not upload the file larger than the Matrix setting(in fact there is no way to stop it). It will throw error during the creation instead.

The default is set to zero which means no size restriction.