Fulfilling Feature Request #1348 Restricting file types that can be uploaded ...
New options have been added to File Section in Global Preferences. All these options can be customized per User Group too.

1. Allow File with No Extension

If this option is set to No, the user can not create any file type of asset with file with no extension. Default is Yes.

2. Allowed File Extensions to Upload

The user can list one or more types of files that are allowed to be used in Matrix. It accepts comma separated extension names such as :

jpg,png,doc,txt,mov

Please note that there is possibility that this given list would conflict with Asset's own restriction in two different ways.

Firstly, the given extension is not allowed in asset itself. For example, Image type of assets do not allow .tiff file to be uploaded, but 'tiff' is listed in the string. During the creation it will print warning message to the users. If the asset does not accept the file type, nothing we can do more.

Secondly, the extension which asset accepts is not listed in the string. In other words, even though Image asset accepts .jpg file, you can restrict the user not be able to use it by omitting it from the list.

Default value is an empty string means no restriction.

3. Allowded File Extensions for Pre-Uploaded files

It's the same as 'Allowed File Extensions to Upload', but it only applies to the pre-uploaded files.

Lastly, if the script runs from the command line, this rule is not applied. Also, Root or Sys Amins are free from this restriction.