The new thesaurus (v.3)
Thesaurus has been updated for fast term import and easier maintenance. Tree ID structure has been dropped. A lot more interface options are available on the
Thesaurus asset's details screen. Now we have support for:

-- Reflexions
Pairs of relations that represent two directions of the same relationship
If a reflexion is created that says that relation XX and YY are relfexive, when term A is linked toterm B using relation XX, term B can automatically be linked
to term A using relation YY, if reflexions are enforced.

-- Hierachy mode
Thesaurus now has a hierarchy mode that can either be enabled or disabled. When enabled, only those terms will appear under the thesaurus in the asset map that
are the absolute parents in a hierarchy relation (configurable). When Hierarchy mode is disabled (unoficially called A to Z mode), all terms are displayed as
children of the thesaurus.

Terms are decoupled from Relations they might be in. This means that it is possible to have a thesaurus where terms are not related at all. Now deleting a term
in the asset map will only delete a particular link that was selected when 'Delete' key was pressed. Only in A-to-Z mode can a term be deleted forever, and
only the term that is the immediat child of the thesaurus.

Thesaurus Term's details screen now has a 'Term Info' panel that allows for a quick navigation of related terms.

XML format has changed slightly. Previous format is still correct. New format supports a <note> tag, which allows for non-relational data to be added to a
term. This allows us to support SC, SN fields in common ANSI-compliant thesauri. Multiple notes are stacked up.

Example XML Format:
<thesaurus>
	<term name="A" >
		<note name="SN">Simple Term Note</note>
		<note name="SC">Category</note>
		<note name="SC">Another Category</note>
		<term name="B" />
	</term>
	<term name="B">
		<relation name="NT">
			<term name="X">
				<term name="Z" />
			</term>
		</relation
	</term>
</thesaurus>

To upgrade run:
step 2
scripts/upgrade_thesaurus_db.php
step 3
compile locale

Because thesaurus is a core asset, every system needs to be upgraded, not just those who had a Thesaurus created.