Additional fix for Bug# 1770
The url of main frame was written in a wrong way, resulting in endlessly nested backend interface. The main frame attempted to load all the frame again. You might observe this if you pass SQ_BACKEND_PAGE=frames in url. For instance, if you access the url like:

http://somewhere.matrix/some/asset/_admin/?SQ_BACKEND_PAGE=frames&asset_ei_screen=metadata

in order to access metadata screen of the asset directly. Also, if the user type the url such as:

http://somewhere.matrix/some/asset/_admin/?asset_ei_screen=metadata

the browser displayed the main frame only with infinite redirect !!!

All fixed.