Bug Fix #1941 current_result_page value is off by 1
The search result index is zero-based internally in search package, i.e. the index of the first result page is zero, and also current_result_page value is correct.

The actual problem for %<term>_query_suggesions% lies in its own keyword replacement. The link added 1 to current_result_page in GET variable. %<term>_query_synonyms% keyword also includes the same problem. Now we get back the consistency that the first result page's index is zero!

The fix made to both of the keywords.