New assets - Page Online Quiz and Online Quiz Question (and derivative(s)).
The Online Quiz assets allows the user to set up a quiz asset that draws from a pool of quiz questions (sorting or grouping them if desired), with the option of sending off results in an email, or simply displaying them to the user. The user details are gathering from the current user, or from a simple_form attached to the quiz.