BUG#1983 ISO8601 and iCal date format keywords not working
The RSS Feed asset will no longer attempt to enforce RFC2822 dates in it's output. This feature relied on PHP's strtotime() function which could not handle dates replaced by the assets ISO8601 or iCal date keywords. This resulted in invalid dates being output in the RSS XML.

Note: The RSS1.0 RSS2.0 standard dictates that dates should be in the RFC2822 format, so you should use the corresponding keyword for this value. Atom feeds should use RFC3339 or ISO8601.