Bug fix #1398 Cannot link LDAP User with space in DN to Matrix group
This bug was reported from Firefox 2.0 and Safari on mac os system for an asset like LDAP user which can contain spaces in an assetid.

When the user select the screen for an asset, the communication between java applet (i.e. assetmap) and the frame breaks if the assetid passed to the main frame contains an empty space since it was not properly url encoded. Consequently many actions from assetmap were blocked or broken, such as not being able to link an asset to somewhere, or bringing details screen up and so on.

Interestingly all the browsers on windows system handled the space without any problems.