Bug Fix for metadata manager.
Previously, when the metadata manager generates a metadata file, it converts all HTML entities in the value array, not in the meta tag.

It caused the problem when the array is used for keyword replacement.

The fix was made to metadata_manager.inc and metadata_field_edit_fns.inc so that $metadata_values array contain the raw value without any conversion.

The value is only converted when it's written for meta tag.