Bug Fix #2203 Global search and replace tool dies on oracle system
Oracle Database Plugin for Search Package generated wrong type of word list. generateWordList() function assumed the passed value is always string, and it wraps it within array type. When the array type is passed in, it nested an array within the array resulting in type matching failure in search manager. This eventually stopped Global Search and Replace Tool from being even started.

3.12.1 and 3.13 have the same cause.

Now the function handles both of string and array type of data. It only puts a string into an array. If an array type is passed, it just returns.