Fix for the previous commit.
-----------------------------------------------------------------------------------------------------
The changes to fix UTF-8, removing rawurlencode() on the changed line, caused an issue where HTML entered into the WYSIWYG (which should be escaped when saved) was being interpreted as unescaped HTML when reloaded into the WYSIWYG - this is a Javascript quirk with TEXTAREA elements where escaped HTML is unescaped when accessed in Javascript, hence the initial fix. Running htmlentities() over the contents when sticking it into the textarea works (including in UTF-8).
-----------------------------------------------------------------------------------------------------

It needs to pass Character Set string as the third argument to htmlentities() function not to break any UTF-8 characters.