Closing slash character added to URL search string.
The problem found in the process of upgrade from 3.10 to 3.12. The old system had file type of asset with the assetid of '90'. It also had an image with the assetid of '9006'.

When the script tries to update asset lookup tables, it forms the search string like '/assets/file/90' and it matches the URL of assetid '9006' too. Consequently some of the assets got the wrong hash number in the lookup table although the file system shows the right hash.

If such a problem found in any system, please run the script again with --fix-lookups option, which finds and fixes wrong asset hash in lookup tables.